requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cfgv==3.2.0
distlib==0.3.1
Django==3.2.1
djangorestframework==3.12.4
filelock==3.0.12
identify==2.2.4
importlib-metadata==4.0.1
importlib-resources==5.1.2
nodeenv==1.6.0
pre-commit==2.12.1
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.4.5
zipp==3.4.1
